chore: address issue #150
